refactor faq section and add responsive
